all normed
