+++ inc/libft/src/ft_put.c
@@ -1,14 +1,14 @@
-/******************************************************************************/
+/* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
 /*   ft_put.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkling <mkling@student.42.fr>              +#+  +:+       +#+        */
+/*   By: ema_blnch <ema_blnch@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/28 18:39:39 by mkling            #+#    #+#             */
-/*   Updated: 2024/12/16 17:17:38 by mkling           ###   ########.fr       */
+/*   Created: 2025/02/27 09:37:14 by ema_blnch         #+#    #+#             */
+/*   Updated: 2025/02/27 09:37:16 by ema_blnch        ###   ########.fr       */
 /*                                                                            */
-/******************************************************************************/
+/* ************************************************************************** */
 
 #include <unistd.h>
 #include "../inc/libft.h"
