+++ src/builtins/export_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   export_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: eblancha <eblancha@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ema_blnch <ema_blnch@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 13:00:46 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/26 10:44:02 by eblancha         ###   ########.fr       */
+/*   Updated: 2025/02/27 09:54:14 by ema_blnch        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,31 +41,6 @@ void	sort_array(char **arr, int len)
 			}
 			j++;
 		}
-	i++;
-	}
-}
-
-char	**lst_to_arr(t_list *env)
-{
-	t_list	*lst;
-	char	**dest;
-	int		i;
-
-	dest = NULL;
-	i = 0;
-	lst = env;
-	dest = (char **)malloc(sizeof(char *) * (len_list(lst) + 1));
-	if (!dest)
-		return (NULL);
-	dest[i] = (lst->str);
-	lst = lst->next;
-	i++;
-	while (lst != env)
-	{
-		dest[i] = (lst->str);
-		lst = lst->next;
 		i++;
 	}
-	dest[i] = NULL;
-	return (dest);
 }
