--- src/execution/old/ast.c
@@ -1,45 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ast.c                                              :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ahamini <ahamini@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/11 10:47:00 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/24 10:24:31 by ahamini          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-/*t_tree	*create_branch(t_shell *shell, int type, void *content)
-{
-	t_tree	*node;
-
-	node = ft_calloc(1, sizeof(t_tree));
-	if (!node)
-		return (set_error(MALLOC_FAIL, shell, "Failed to malloc tree"), NULL);
-	node->type = type;
-	node->content = content;
-	node->left = NULL;
-	node->left = NULL;
-	return (node);
-}
-
-void	command_into_array(t_cmd *cmd)
-{
-	int	argc;
-	int	index;
-
-	argc = ft_lstsize(cmd->arg_list);
-	cmd->argv = ft_calloc(sizeof(char *), argc + 1);
-	if (!cmd->argv)
-		return (set_cmd_error(MALLOC_FAIL, cmd, "Failed to malloc argv"));
-	index = 0;
-	while (index < argc)
-	{
-		cmd->argv[index] = cmd->arg_list->content;
-		cmd->arg_list = cmd->arg_list->next;
-		index++;
-	}
-}*/
