--- src/execution/old/errors.c
@@ -1,52 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   errors.c                                           :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ahamini <ahamini@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/11 10:47:12 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/24 10:24:05 by ahamini          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-/*void	print_error(void)
-{
-	ft_putstr_fd(SHELL_NAME, STDERR_FILENO);
-	ft_putstr_fd(": ", STDERR_FILENO);
-}
-
-void	set_error(int err_code, t_shell *shell, char *err_message)
-{
-	print_error();
-	ft_putstr_fd(err_message, STDERR_FILENO);
-	ft_putstr_fd("\n", STDERR_FILENO);
-	shell->critical_er = err_code;
-}
-
-void	set_cmd_error(int err_code, t_cmd *cmd, char *err_message)
-{
-	print_error();
-	if (cmd->arg_list && (char *)cmd->arg_list->content)
-	{
-		ft_putstr_fd((char *)cmd->arg_list->content, STDERR_FILENO);
-		ft_putstr_fd(": ", STDERR_FILENO);
-	}
-	ft_putstr_fd(err_message, STDERR_FILENO);
-	ft_putstr_fd("\n", STDERR_FILENO);
-	cmd->exit_code = err_code;
-}
-
-void	set_error_if(int condition, int err_code, t_shell *shell,
-		char *err_message)
-{
-	if (condition)
-		set_error(err_code, shell, err_message);
-}
-
-int	catch_error(t_shell *shell)
-{
-	return (shell->critical_er);
-}*/
