--- src/execution/old/heredoc.c
@@ -1,60 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   heredoc.c                                          :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ahamini <ahamini@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/11 10:47:30 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/24 10:11:16 by ahamini          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-static bool	read_in_stdin(t_shell *shell, int fd, char *word)
-{
-	char	*buf;
-
-	while (1)
-	{
-		buf = NULL;
-		buf = readline("> ");
-		if (!buf)
-		{
-			print_error2("warning: here-document delimited by end-of-file ");
-			print_error2("(wanted '");
-			print_error2(word);
-			print_error2("')\n");
-			break ;
-		}
-		if (!ft_strncmp(word, buf, INT_MAX))
-			break ;
-		if (!replace_dollar(&buf, shell))
-			free_all(shell, ERR_MALLOC, EXT_MALLOC);
-		write(fd, buf, ft_strlen(buf));
-		write(fd, "\n", 1);
-		free(buf);
-	}
-	free(buf);
-	close(fd);
-	return (true);
-}
-
-int	here_doc(t_shell *shell, char *word)
-{
-	int	fd;
-
-	fd = open(".heredoc.tmp", O_CREAT | O_WRONLY | O_TRUNC, 0644);
-	if (fd < 0)
-		return (-1);
-	if (!read_in_stdin(shell, fd, word))
-	{
-		unlink(".heredoc.tmp");
-		return (-1);
-	}
-	fd = open(".heredoc.tmp", O_RDONLY);
-	if (fd > 0)
-		unlink(".heredoc.tmp");
-	return (fd);
-}
