+++ src/parsing/cmd_fd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cmd_fd.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ahamini <ahamini@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ema_blnch <ema_blnch@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 09:44:21 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/24 09:45:34 by ahamini          ###   ########.fr       */
+/*   Updated: 2025/02/27 09:51:37 by ema_blnch        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
