+++ src/parsing/cmd_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cmd_utils.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ahamini <ahamini@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ema_blnch <ema_blnch@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 10:17:56 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/24 10:18:26 by ahamini          ###   ########.fr       */
+/*   Updated: 2025/02/27 09:50:14 by ema_blnch        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
