+++ src/parsing/errors.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   errors.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ahamini <ahamini@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ema_blnch <ema_blnch@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 09:46:58 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/24 09:47:58 by ahamini          ###   ########.fr       */
+/*   Updated: 2025/02/27 09:50:18 by ema_blnch        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
