+++ src/parsing/quote.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   quote.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ahamini <ahamini@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ema_blnch <ema_blnch@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 09:01:57 by ahamini           #+#    #+#             */
-/*   Updated: 2025/02/24 14:26:33 by ahamini          ###   ########.fr       */
+/*   Updated: 2025/02/27 09:51:38 by ema_blnch        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
